Fix kube-update GO version detection
GO version was not updated on the latest Kube patch update because
the local repo was not on the expected GO version for the prior Kube version.
Update the logic so that current GO version is determined by what is
actually being used by the repo and set the current GO version of the repo
to what it should be using for the latest Kube release.